Option2/Lab/src/tts-server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-server"
version = "0.1.0"
description = "XTTS-v2 TTS inference server for GPU workload on ACA"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "torch>=2.1.0",
    "TTS==0.22.0",
    "numpy",
    "huggingface-hub>=0.20",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27"]

[tool.setuptools]
py-modules = ["main", "model_loader"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
